+++ mail_activity_done/README.rst
@@ -14,13 +14,13 @@ Mail Activity Done
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsocial-lightgray.png?logo=github
-    :target: https://github.com/OCA/social/tree/14.0/mail_activity_done
+    :target: https://github.com/OCA/social/tree/15.0/mail_activity_done
     :alt: OCA/social
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/social-14-0/social-14-0-mail_activity_done
+    :target: https://translation.odoo-community.org/projects/social-15-0/social-15-0-mail_activity_done
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/205/14.0
+    :target: https://runbot.odoo-community.org/runbot/205/15.0
     :alt: Try me on Runbot
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
@@ -28,12 +28,23 @@ Mail Activity Done
 This module implements the capability to keep activities that have been
 completed, for future reporting, by setting them with the boolean 'Done'.
 
-The activities that have been completed will not appear in the chatter.
-
 **Table of contents**
 
 .. contents::
    :local:
+
+Usage
+=====
+
+To use this module, you need to:
+
+#. Access to an activity from the systray activities menu.
+#. Once finished, open the activity and mark it as Done.
+
+To check activities:
+
+#. Go to Settings -> Technical -> Activities
+#. To see finished activities filter by Completed Activities
 
 Bug Tracker
 ===========
@@ -41,7 +52,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/social/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_activity_done%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_activity_done%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -60,6 +71,7 @@ Contributors
 * Eduardo Magdalena <emagdalena@c2i.es> (C2i Change 2 improve http://www.c2i.es)
 * Radovan Skolnik <radovan@skolnik.info> (https://www.kema.sk)
 * Manuel Regidor <manuel.regidor@sygel.es> (https://www.sygel.es)
+* Bernat Puig <bernat.puig@forgeflow.com> (www.forgeflow.com)
 
 Maintainers
 ~~~~~~~~~~~
@@ -74,6 +86,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/social <https://github.com/OCA/social/tree/14.0/mail_activity_done>`_ project on GitHub.
+This module is part of the `OCA/social <https://github.com/OCA/social/tree/15.0/mail_activity_done>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
