+++ mail_activity_done/i18n/mail_activity_done.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -15,7 +15,7 @@ msgstr ""
 
 #. module: mail_activity_done
 #. openerp-web
-#: code:addons/mail_activity_done/static/src/js/mail_activity.js:0
+#: code:addons/mail_activity_done/static/src/js/mail_activity.esm.js:0
 #, python-format
 msgid "%d days overdue"
 msgstr ""
@@ -26,16 +26,8 @@ msgid "Active"
 msgstr ""
 
 #. module: mail_activity_done
-#: model:ir.model.fields,field_description:mail_activity_done.field_account_bank_statement_line__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_account_journal__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_account_move__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_account_payment__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_event_event__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_event_registration__activity_ids
 #: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity_mixin__activity_ids
 #: model:ir.model.fields,field_description:mail_activity_done.field_mailing_mailing__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_product_product__activity_ids
-#: model:ir.model.fields,field_description:mail_activity_done.field_product_template__activity_ids
 #: model:ir.model.fields,field_description:mail_activity_done.field_res_partner__activity_ids
 #: model:ir.model.fields,field_description:mail_activity_done.field_res_users__activity_ids
 msgid "Activities"
@@ -63,13 +55,6 @@ msgid "Completed Date"
 msgstr ""
 
 #. module: mail_activity_done
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity__display_name
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity_mixin__display_name
-#: model:ir.model.fields,field_description:mail_activity_done.field_res_users__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: mail_activity_done
 #: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity__done
 #: model:ir.model.fields.selection,name:mail_activity_done.selection__mail_activity__state__done
 msgid "Done"
@@ -77,23 +62,9 @@ msgstr ""
 
 #. module: mail_activity_done
 #. openerp-web
-#: code:addons/mail_activity_done/static/src/js/mail_activity.js:0
+#: code:addons/mail_activity_done/static/src/js/mail_activity.esm.js:0
 #, python-format
 msgid "Due in %d days"
-msgstr ""
-
-#. module: mail_activity_done
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity__id
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity_mixin__id
-#: model:ir.model.fields,field_description:mail_activity_done.field_res_users__id
-msgid "ID"
-msgstr ""
-
-#. module: mail_activity_done
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity____last_update
-#: model:ir.model.fields,field_description:mail_activity_done.field_mail_activity_mixin____last_update
-#: model:ir.model.fields,field_description:mail_activity_done.field_res_users____last_update
-msgid "Last Modified on"
 msgstr ""
 
 #. module: mail_activity_done
@@ -108,14 +79,14 @@ msgstr ""
 
 #. module: mail_activity_done
 #. openerp-web
-#: code:addons/mail_activity_done/static/src/js/mail_activity.js:0
+#: code:addons/mail_activity_done/static/src/js/mail_activity.esm.js:0
 #, python-format
 msgid "Today"
 msgstr ""
 
 #. module: mail_activity_done
 #. openerp-web
-#: code:addons/mail_activity_done/static/src/js/mail_activity.js:0
+#: code:addons/mail_activity_done/static/src/js/mail_activity.esm.js:0
 #, python-format
 msgid "Tomorrow"
 msgstr ""
@@ -127,7 +98,7 @@ msgstr ""
 
 #. module: mail_activity_done
 #. openerp-web
-#: code:addons/mail_activity_done/static/src/js/mail_activity.js:0
+#: code:addons/mail_activity_done/static/src/js/mail_activity.esm.js:0
 #, python-format
 msgid "Yesterday"
 msgstr ""
