+++ mail_activity_done/readme/DESCRIPTION.rst
@@ -1,4 +1,2 @@
 This module implements the capability to keep activities that have been
 completed, for future reporting, by setting them with the boolean 'Done'.
-
-The activities that have been completed will not appear in the chatter.
